CRS-2805: Fix app traces
The logger being initialised in azureAppInsights was causing the bunyan logs not to be written to AppTraces in app insights.

Removed the log as per the template project.

Also moved the custom dimensions to ensure that any express instrumentation also happens.
